pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pronote-mcp"
version = "0.1.0"
description = "MCP server for Pronote (read-only) via Monlycée ENT"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Thomas Greissler" }]
keywords = ["mcp", "pronote", "monlycee", "ent", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "mcp[cli]>=1.25,<2",
    "pronotepy>=2.14",
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
pronote-mcp = "pronote_mcp.server:main"
pronote-mcp-http = "pronote_mcp.server_http:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/pronote-mcp"
Issues = "https://github.com/YOUR_USERNAME/pronote-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pronote_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
